feat(core): AssemblyInfo support and new 'merge-changelog' and 'revision' CLI options

+++ Enterwell.CI.Changelog/ChangelogManagerCommand.cs
@@ -61,6 +61,22 @@ namespace Enterwell.CI.Changelog
         public (bool isEnabled, string projectFilePath) SetVersion { get; set; }
 
         /// <summary>
+        /// Optional parameter to the CLI application representing if the application should support 4 number long versions.
+        /// </summary>
+        [Option("-r|--revision",
+            CommandOptionType.SingleValue,
+            Description = "If the revision number is provided, 4 number long versions will be supported.\nThis is write-only, which means the revision number will not be bumped automatically, but only replaced if provided.")]
+        public int? RevisionNumber { get; set; }
+
+        /// <summary>
+        /// Optional parameter to the CLI application representing should the newly generated changelog section be merged to the changelog. Defaults to true.
+        /// </summary>
+        [Option("-mc|--merge-changelog",
+            CommandOptionType.SingleValue,
+            Description = "Should the newly generated changelog section be merged to the changelog. If set to false, the merge step is skipped.")]
+        public bool ShouldMergeChangelog { get; set; } = true;
+
+        /// <summary>
         /// Main method of the application. The method delegates all the work to the appropriate services.
         /// </summary>
         /// <returns>An asynchronous task.</returns>
@@ -76,18 +92,21 @@ namespace Enterwell.CI.Changelog
                 // If the set-version flag is set
                 if (this.SetVersion.isEnabled)
                 {
-                    await this.fileWriterService.BumpProjectFileVersion(versionInformation.SemanticVersion, this.SetVersion.projectFilePath);
+                    await this.fileWriterService.BumpProjectFileVersion(versionInformation.SemanticVersion, this.SetVersion.projectFilePath, this.RevisionNumber);
                 }
 
                 // Building the string representing the new changelog section for the new bumped version
                 var newChangelogSection = this.markdownTextService.BuildChangelogSection(versionInformation);
                 var elementToInsertChangelogSectionBefore = this.markdownTextService.ToH2(string.Empty);
 
-                // Write the newly built section to the changelog
-                await this.fileWriterService.WriteToChangelog(newChangelogSection, this.ChangelogLocation, elementToInsertChangelogSectionBefore);
+                if (this.ShouldMergeChangelog)
+                {
+                    // Write the newly built section to the changelog
+                    await this.fileWriterService.WriteToChangelog(newChangelogSection, this.ChangelogLocation, elementToInsertChangelogSectionBefore);
 
-                // Delete the accepted change files
-                this.changeGatheringService.RemoveAcceptedChanges(versionInformation.Changes);
+                    // Delete the accepted change files
+                    this.changeGatheringService.RemoveAcceptedChanges(versionInformation.Changes);
+                }
 
                 logger.LogSuccess(newChangelogSection);
             }
